Split p2p manager into Dialer and Listener
