+++ core/connection/manager_test.go
@@ -26,11 +26,9 @@ import (
 	"github.com/mysteriumnetwork/node/communication"
 	"github.com/mysteriumnetwork/node/communication/nats"
 	"github.com/mysteriumnetwork/node/core/ip"
-	"github.com/mysteriumnetwork/node/eventbus"
 	"github.com/mysteriumnetwork/node/identity"
 	"github.com/mysteriumnetwork/node/market"
 	"github.com/mysteriumnetwork/node/money"
-	"github.com/mysteriumnetwork/node/nat/traversal"
 	"github.com/mysteriumnetwork/node/p2p"
 	"github.com/mysteriumnetwork/node/session"
 	"github.com/mysteriumnetwork/node/session/connectivity"
@@ -120,21 +118,11 @@ func (tc *testContext) SetupTest() {
 
 	brokerConn := nats.StartConnectionMock()
 	brokerConn.MockResponse("fake-node-1.p2p-config-exchange", []byte("123"))
-	mockBroker := &mockBroker{conn: brokerConn}
-
-	pinger := traversal.NewPinger(&traversal.PingConfig{
-		Interval: 1 * time.Second,
-		Timeout:  3 * time.Second,
-	}, eventbus.New())
-
-	p2pManager := p2p.NewManager(mockBroker, "", func(_ identity.Identity) identity.Signer {
-		return &identity.SignerFake{}
-	}, ip.NewResolverMock("127.0.0.1"), pinger, pinger)
 
 	tc.connManager = NewManager(
 		dialogCreator,
 		func(paymentInfo session.PaymentInfo,
-			dialog communication.Dialog, channel *p2p.Channel,
+			dialog communication.Dialog, channel p2p.Channel,
 			consumer, provider, accountant identity.Identity, proposal market.ServiceProposal, sessionID string) (PaymentIssuer, error) {
 			tc.MockPaymentIssuer = &MockPaymentIssuer{
 				initialState:      paymentInfo,
@@ -150,7 +138,7 @@ func (tc *testContext) SetupTest() {
 		tc.ipCheckParams,
 		tc.statsReportInterval,
 		func(ID identity.Identity) uint64 { return 0 },
-		p2pManager,
+		&mockP2PDialer{},
 	)
 }
 
@@ -599,10 +587,19 @@ func (mpm *mockPaymentMethod) GetRate() market.PaymentRate {
 	return mpm.rate
 }
 
-type mockBroker struct {
-	conn nats.Connection
+type mockP2PDialer struct {
 }
 
-func (m *mockBroker) Connect(serverURIs ...string) (nats.Connection, error) {
-	return m.conn, nil
+func (m mockP2PDialer) Dial(consumerID, providerID identity.Identity, serviceType string, timeout time.Duration) (p2p.Channel, error) {
+	return &mockP2PChannel{}, nil
+}
+
+type mockP2PChannel struct {
+}
+
+func (m mockP2PChannel) Send(topic string, msg *p2p.Message) (*p2p.Message, error) {
+	return nil, nil
+}
+
+func (m mockP2PChannel) Handle(topic string, handler p2p.HandlerFunc) {
 }
