+++ core/service/manager.go
@@ -76,8 +76,8 @@ func NewManager(
 	discoveryFactory DiscoveryFactory,
 	eventPublisher Publisher,
 	policyOracle *policy.Oracle,
-	p2pManager *p2p.Manager,
-	sessionManager func(proposal market.ServiceProposal, serviceID string, channel *p2p.Channel) *session.Manager,
+	p2pListener p2p.Listener,
+	sessionManager func(proposal market.ServiceProposal, serviceID string, channel p2p.Channel) *session.Manager,
 ) *Manager {
 	return &Manager{
 		serviceRegistry:      serviceRegistry,
@@ -87,7 +87,7 @@ func NewManager(
 		discoveryFactory:     discoveryFactory,
 		eventPublisher:       eventPublisher,
 		policyOracle:         policyOracle,
-		p2pManager:           p2pManager,
+		p2pManager:           p2pListener,
 		sessionManager:       sessionManager,
 	}
 }
@@ -104,8 +104,8 @@ type Manager struct {
 	eventPublisher   Publisher
 	policyOracle     *policy.Oracle
 
-	p2pManager     *p2p.Manager
-	sessionManager func(proposal market.ServiceProposal, serviceID string, channel *p2p.Channel) *session.Manager
+	p2pManager     p2p.Listener
+	sessionManager func(proposal market.ServiceProposal, serviceID string, channel p2p.Channel) *session.Manager
 }
 
 // Start starts an instance of the given service type if knows one in service registry.
@@ -146,7 +146,7 @@ func (manager *Manager) Start(providerID identity.Identity, serviceType string, 
 		return id, err
 	}
 
-	err = manager.p2pManager.SubscribeChannel(providerID, func(ch *p2p.Channel) {
+	err = manager.p2pManager.Listen(providerID, serviceType, func(ch p2p.Channel) {
 		mng := manager.sessionManager(proposal, string(id), ch)
 		subscribeSessionCreate(mng, ch, service, id)
 		subscribeSessionAcknowledge(mng, ch)
